Fixed issues caused during renaming
