+++ EvoWPF/Evo.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.32112.339
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WaistlandGameWPF", "WaistlandGameWPF\WaistlandGameWPF.csproj", "{D689C9E5-66BA-4552-8B06-4ED6EDD50BDF}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EvoWPF", "Evo\Evo.csproj", "{D689C9E5-66BA-4552-8B06-4ED6EDD50BDF}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
