+++ EvoWPF/Evo/App.xaml
@@ -1,7 +1,7 @@
-﻿<Application x:Class="WaistlandGameWPF.App"
+﻿<Application x:Class="Evo.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:WaistlandGameWPF"
+             xmlns:local="clr-namespace:Evo"
              StartupUri="MainWindow.xaml">
     <Application.Resources>
