+++ EvoWPF/Evo/GameObjects/Animal.cs
@@ -2,7 +2,7 @@
 
 
 //later: add entity levels wich affect how skilled they are ex: high lvl might have high accuracy
-namespace WaistlandGameWPF
+namespace Evo
 {
 [System.Serializable]
 public class Animal : Entity {
