+++ EvoWPF/Evo/GameObjects/Dialogue.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace WaistlandGameWPF
+namespace Evo
 {
     [System.Serializable]
     class Dialogue
