+++ EvoWPF/Evo/GameObjects/GameObjectManager.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace WaistlandGameWPF.WastelandGame.GameObjects
+namespace Evo
 {
 	//package wasteland.gameObject;
