+++ EvoWPF/Evo/GameObjects/Hitbox.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Diagnostics;
 
-namespace WaistlandGameWPF
+namespace Evo
 {
 	/*
     package wasteland.entity;
