+++ EvoWPF/Evo/GameObjects/Material.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace WaistlandGameWPF
+namespace Evo
 {
 	//package wasteland.entity;
 	[System.Serializable]
