+++ EvoWPF/Evo/MainWindow.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="WaistlandGameWPF.MainWindow"
+﻿<Window x:Class="Evo.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:WaistlandGameWPF"
+        xmlns:local="clr-namespace:Evo"
         mc:Ignorable="d"
         Title="MainWindow" Height="720" Width="1280">
     <Grid Margin="0,0,0,-6">
